page 계산에 사용하는 ceil method 사용시 double 타입으로 안해주면 오동작? 하는 버그 수정, 한페이지 이내 갯수일경우 발생

+++ src/main/java/itwise/coreasset/shaman/api/sample/BuildProfileApi.java
@@ -225,7 +225,7 @@ public class BuildProfileApi {
 		) {
 		
 		int count = profileMapper.count();
-		int lastPage = (int)Math.ceil(count / limit);
+		int lastPage = (int)Math.ceil((double)count / limit);
 		
 		if(page < 1) page = 1;
 		else if(page > lastPage) page = lastPage;
