+++ src/main/java/itwise/coreasset/shaman/api/sample/ProjectGroupApi.java
@@ -198,7 +198,7 @@ public class ProjectGroupApi {
 		) {
 		
 		int count = groupMapper.count();
-		int lastPage = (int)Math.ceil(count / limit) + 1;
+		int lastPage = (int)Math.ceil((double)count / limit) + 1;
 		
 		if(page < 1) page = 1;
 		else if(page > lastPage) page = lastPage;
